--- src/app/create/_components/ImageUploader.css.ts
@@ -1,9 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const label = style({
-  cursor: 'pointer',
-});
-
-export const input = style({
-  display: 'none',
-});
